- Clear old markers on reload - Display distance in meters only for <100m

+++ app/src/main/java/com/helphelp2/android/models/Place.java
@@ -50,7 +50,7 @@ public class Place {
 
         if (dist != 0) {
             if (dist < 100) {
-                res += dist + "m";
+                res += (int)dist + "m";
             } else if (dist >= 100 && dist <= 1000) {
                 res += (int) (dist / 100) + "00m";
             } else {
